// src/ipc/dtos/enums/InferenceChiSquareCategoriesEnum.ts
export enum InferenceChiSquareCategoriesEnum {
    histologicalTypes = 'histologicalTypes',
    tClassification = 'tClassification',
    nClassification = 'nClassification',
    mClassification = 'mClassification',
    persistence = 'persistence',
    recurrence = 'recurrence',
    state = 'state',
}